Extract DataAnalysisOperation from AbstractAggregate
